Simple server address mode & bug fixes
- Added a new default server address input mode, simple mode, where the
full address is input by the user as a single String. The user can
select the old input mode by unchecking a CheckBox in ServerEditor. The
mode used to save the ServerConnection is saved along with the server
information, so that everywhere the server address is read, which field
to use can be determined by the address mode.
- Added application-wide theme without title bar, except in main menu
which keeps a title bar.
- Made the (rotating) web connection icon opaque to be seen more clearly
against black backgrounds.
- Changed the name of the styles file from style.xml to styles.xml and
removed an unused theme.
- Added an alert stopping the user from trying to upload lines or
polygons when there are no point sequences stored.
- Uploading a line or polygon layer no longer removes points that are
not included in point sequences.
- Setting the BackboneSvc-monitored connection state and the row being
connected to is now done at the same time.
- Changing the text of the Upload button in LayerViewer when enabling or
disabling is now optional.
- ServerEditor no longer stores references to its Views, making view
recycling more efficient, hopefully.
- User input in ServerEditor is now sent through onSaveInstanceState
instead of storing it on the BackboneSvc.
- Increased time-out limit for DescribeFeatureType requests from 15 to
10.